Fix clang-tidy bugprone-* warnings;

+++ mlir/lib/Conversion/LoSPNtoCPU/NodePatterns.cpp
@@ -425,7 +425,7 @@ mlir::LogicalResult mlir::spn::HistogramLowering::matchAndRewrite(mlir::spn::low
       indexVal = (computesLog) ? log(values[i]) : values[i];
     } else {
       // Fill up with 0 if no value was defined by the histogram.
-      indexVal = (computesLog) ? -INFINITY : 0;
+      indexVal = (computesLog) ? static_cast<double>(-INFINITY) : 0;
     }
     // Construct attribute with constant value. Need to distinguish cases here due to different builder methods.
     if (resultType.isIntOrIndex()) {
